Crate rusoto_appsync

AWS AppSync provides API actions for creating and interacting with data sources using GraphQL from your application.

If you're using the service, you're probably looking for AppSyncClient and AppSync.

Structs

ApiKey

Describes an API key.

Customers invoke AWS AppSync GraphQL APIs with API keys as an identity mechanism. There are two key versions:

da1: This version was introduced at launch in November 2017. These keys always expire after 7 days. Key expiration is managed by DynamoDB TTL. The keys will cease to be valid after Feb 21, 2018 and should not be used after that date.

ListApiKeys returns the expiration time in milliseconds.

CreateApiKey returns the expiration time in milliseconds.

UpdateApiKey is not available for this key version.

DeleteApiKey deletes the item from the table.

Expiration is stored in DynamoDB as milliseconds. This results in a bug where keys are not automatically deleted because DynamoDB expects the TTL to be stored in seconds. As a one-time action, we will delete these keys from the table after Feb 21, 2018.

da2: This version was introduced in February 2018 when AppSync added support to extend key expiration.

ListApiKeys returns the expiration time in seconds.

CreateApiKey returns the expiration time in seconds and accepts a user-provided expiration time in seconds.

UpdateApiKey returns the expiration time in seconds and accepts a user-provided expiration time in seconds. Key expiration can only be updated while the key has not expired.

DeleteApiKey deletes the item from the table.

Expiration is stored in DynamoDB as seconds.

AppSyncClient
A client for the AWSAppSync API.

CreateApiKeyRequest

CreateApiKeyResponse

CreateDataSourceRequest

CreateDataSourceResponse

CreateGraphqlApiRequest

CreateGraphqlApiResponse

CreateResolverRequest

CreateResolverResponse

CreateTypeRequest

CreateTypeResponse

DataSource
Describes a data source.

DeleteApiKeyRequest

DeleteApiKeyResponse

DeleteDataSourceRequest

DeleteDataSourceResponse

DeleteGraphqlApiRequest

DeleteGraphqlApiResponse

DeleteResolverRequest

DeleteResolverResponse

DeleteTypeRequest

DeleteTypeResponse

DynamodbDataSourceConfig
Describes a DynamoDB data source configuration.

ElasticsearchDataSourceConfig
Describes an Elasticsearch data source configuration.

GetDataSourceRequest

GetDataSourceResponse

GetGraphqlApiRequest

GetGraphqlApiResponse

GetIntrospectionSchemaRequest

GetIntrospectionSchemaResponse

GetResolverRequest

GetResolverResponse

GetSchemaCreationStatusRequest

GetSchemaCreationStatusResponse

GetTypeRequest

GetTypeResponse

GraphqlApi
Describes a GraphQL API.

HttpDataSourceConfig
Describes a Http data source configuration.

LambdaDataSourceConfig
Describes a Lambda data source configuration.

ListApiKeysRequest

ListApiKeysResponse

ListDataSourcesRequest

ListDataSourcesResponse

ListGraphqlApisRequest

ListGraphqlApisResponse

ListResolversRequest

ListResolversResponse

ListTypesRequest

ListTypesResponse

LogConfig
The CloudWatch Logs configuration.

OpenIDConnectConfig
Describes an Open Id Connect configuration.

Resolver
Describes a resolver.

StartSchemaCreationRequest

StartSchemaCreationResponse

Type
Describes a type.

UpdateApiKeyRequest

UpdateApiKeyResponse

UpdateDataSourceRequest

UpdateDataSourceResponse

UpdateGraphqlApiRequest

UpdateGraphqlApiResponse

UpdateResolverRequest

UpdateResolverResponse

UpdateTypeRequest

UpdateTypeResponse

UserPoolConfig
Describes an Amazon Cognito User Pool configuration.

Enums

CreateApiKeyError
Errors returned by CreateApiKey

CreateDataSourceError
Errors returned by CreateDataSource

CreateGraphqlApiError
Errors returned by CreateGraphqlApi

CreateResolverError
Errors returned by CreateResolver

CreateTypeError
Errors returned by CreateType

DeleteApiKeyError
Errors returned by DeleteApiKey

DeleteDataSourceError
Errors returned by DeleteDataSource

DeleteGraphqlApiError
Errors returned by DeleteGraphqlApi

DeleteResolverError
Errors returned by DeleteResolver

DeleteTypeError
Errors returned by DeleteType

GetDataSourceError
Errors returned by GetDataSource

GetGraphqlApiError
Errors returned by GetGraphqlApi

GetIntrospectionSchemaError
Errors returned by GetIntrospectionSchema

GetResolverError
Errors returned by GetResolver

GetSchemaCreationStatusError
Errors returned by GetSchemaCreationStatus

GetTypeError
Errors returned by GetType

ListApiKeysError
Errors returned by ListApiKeys

ListDataSourcesError
Errors returned by ListDataSources

ListGraphqlApisError
Errors returned by ListGraphqlApis

ListResolversError
Errors returned by ListResolvers

ListTypesError
Errors returned by ListTypes

StartSchemaCreationError
Errors returned by StartSchemaCreation

UpdateApiKeyError
Errors returned by UpdateApiKey

UpdateDataSourceError
Errors returned by UpdateDataSource

UpdateGraphqlApiError
Errors returned by UpdateGraphqlApi

UpdateResolverError
Errors returned by UpdateResolver

UpdateTypeError
Errors returned by UpdateType

Traits

AppSync
Trait representing the capabilities of the AWSAppSync API. AWSAppSync clients implement this trait.