Trait rusoto_appsync::AppSync

pub trait AppSync {
    fn create_api_key(
        &self, 
        input: CreateApiKeyRequest
    ) -> RusotoFuture<CreateApiKeyResponse, CreateApiKeyError>;

    fn create_data_source(
        &self, 
        input: CreateDataSourceRequest
    ) -> RusotoFuture<CreateDataSourceResponse, CreateDataSourceError>;

    fn create_graphql_api(
        &self, 
        input: CreateGraphqlApiRequest
    ) -> RusotoFuture<CreateGraphqlApiResponse, CreateGraphqlApiError>;

    fn create_resolver(
        &self, 
        input: CreateResolverRequest
    ) -> RusotoFuture<CreateResolverResponse, CreateResolverError>;

    fn create_type(
        &self, 
        input: CreateTypeRequest
    ) -> RusotoFuture<CreateTypeResponse, CreateTypeError>;

    fn delete_api_key(
        &self, 
        input: DeleteApiKeyRequest
    ) -> RusotoFuture<DeleteApiKeyResponse, DeleteApiKeyError>;

    fn delete_data_source(
        &self, 
        input: DeleteDataSourceRequest
    ) -> RusotoFuture<DeleteDataSourceResponse, DeleteDataSourceError>;

    fn delete_graphql_api(
        &self, 
        input: DeleteGraphqlApiRequest
    ) -> RusotoFuture<DeleteGraphqlApiResponse, DeleteGraphqlApiError>;

    fn delete_resolver(
        &self, 
        input: DeleteResolverRequest
    ) -> RusotoFuture<DeleteResolverResponse, DeleteResolverError>;

    fn delete_type(
        &self, 
        input: DeleteTypeRequest
    ) -> RusotoFuture<DeleteTypeResponse, DeleteTypeError>;

    fn get_data_source(
        &self, 
        input: GetDataSourceRequest
    ) -> RusotoFuture<GetDataSourceResponse, GetDataSourceError>;

    fn get_graphql_api(
        &self, 
        input: GetGraphqlApiRequest
    ) -> RusotoFuture<GetGraphqlApiResponse, GetGraphqlApiError>;

    fn get_introspection_schema(
        &self, 
        input: GetIntrospectionSchemaRequest
    ) -> RusotoFuture<GetIntrospectionSchemaResponse, GetIntrospectionSchemaError>;

    fn get_resolver(
        &self, 
        input: GetResolverRequest
    ) -> RusotoFuture<GetResolverResponse, GetResolverError>;

    fn get_schema_creation_status(
        &self, 
        input: GetSchemaCreationStatusRequest
    ) -> RusotoFuture<GetSchemaCreationStatusResponse, GetSchemaCreationStatusError>;

    fn get_type(
        &self, 
        input: GetTypeRequest
    ) -> RusotoFuture<GetTypeResponse, GetTypeError>;

    fn list_api_keys(
        &self, 
        input: ListApiKeysRequest
    ) -> RusotoFuture<ListApiKeysResponse, ListApiKeysError>;

    fn list_data_sources(
        &self, 
        input: ListDataSourcesRequest
    ) -> RusotoFuture<ListDataSourcesResponse, ListDataSourcesError>;

    fn list_graphql_apis(
        &self, 
        input: ListGraphqlApisRequest
    ) -> RusotoFuture<ListGraphqlApisResponse, ListGraphqlApisError>;

    fn list_resolvers(
        &self, 
        input: ListResolversRequest
    ) -> RusotoFuture<ListResolversResponse, ListResolversError>;

    fn list_types(
        &self, 
        input: ListTypesRequest
    ) -> RusotoFuture<ListTypesResponse, ListTypesError>;

    fn start_schema_creation(
        &self, 
        input: StartSchemaCreationRequest
    ) -> RusotoFuture<StartSchemaCreationResponse, StartSchemaCreationError>;

    fn update_api_key(
        &self, 
        input: UpdateApiKeyRequest
    ) -> RusotoFuture<UpdateApiKeyResponse, UpdateApiKeyError>;

    fn update_data_source(
        &self, 
        input: UpdateDataSourceRequest
    ) -> RusotoFuture<UpdateDataSourceResponse, UpdateDataSourceError>;

    fn update_graphql_api(
        &self, 
        input: UpdateGraphqlApiRequest
    ) -> RusotoFuture<UpdateGraphqlApiResponse, UpdateGraphqlApiError>;

    fn update_resolver(
        &self, 
        input: UpdateResolverRequest
    ) -> RusotoFuture<UpdateResolverResponse, UpdateResolverError>;

    fn update_type(
        &self, 
        input: UpdateTypeRequest
    ) -> RusotoFuture<UpdateTypeResponse, UpdateTypeError>;
}

Trait representing the capabilities of the AWSAppSync API. AWSAppSync clients implement this trait.

Required Methods

fn create_api_key(
&self,
input: CreateApiKeyRequest
) -> RusotoFuture<CreateApiKeyResponse, CreateApiKeyError>

Creates a unique key that you can distribute to clients who are executing your API.

fn create_data_source(
&self,
input: CreateDataSourceRequest
) -> RusotoFuture<CreateDataSourceResponse, CreateDataSourceError>

Creates a DataSource object.

fn create_graphql_api(
&self,
input: CreateGraphqlApiRequest
) -> RusotoFuture<CreateGraphqlApiResponse, CreateGraphqlApiError>

Creates a GraphqlApi object.

fn create_resolver(
&self,
input: CreateResolverRequest
) -> RusotoFuture<CreateResolverResponse, CreateResolverError>

Creates a Resolver object.

A resolver converts incoming requests into a format that a data source can understand and converts the data source's responses into GraphQL.

fn create_type(
&self,
input: CreateTypeRequest
) -> RusotoFuture<CreateTypeResponse, CreateTypeError>

Creates a Type object.

fn delete_api_key(
&self,
input: DeleteApiKeyRequest
) -> RusotoFuture<DeleteApiKeyResponse, DeleteApiKeyError>

Deletes an API key.

fn delete_data_source(
&self,
input: DeleteDataSourceRequest
) -> RusotoFuture<DeleteDataSourceResponse, DeleteDataSourceError>

Deletes a DataSource object.

fn delete_graphql_api(
&self,
input: DeleteGraphqlApiRequest
) -> RusotoFuture<DeleteGraphqlApiResponse, DeleteGraphqlApiError>

Deletes a GraphqlApi object.

fn delete_resolver(
&self,
input: DeleteResolverRequest
) -> RusotoFuture<DeleteResolverResponse, DeleteResolverError>

Deletes a Resolver object.

fn delete_type(
&self,
input: DeleteTypeRequest
) -> RusotoFuture<DeleteTypeResponse, DeleteTypeError>

Deletes a Type object.

fn get_data_source(
&self,
input: GetDataSourceRequest
) -> RusotoFuture<GetDataSourceResponse, GetDataSourceError>

Retrieves a DataSource object.

fn get_graphql_api(
&self,
input: GetGraphqlApiRequest
) -> RusotoFuture<GetGraphqlApiResponse, GetGraphqlApiError>

Retrieves a GraphqlApi object.

fn get_introspection_schema(
&self,
input: GetIntrospectionSchemaRequest
) -> RusotoFuture<GetIntrospectionSchemaResponse, GetIntrospectionSchemaError>

Retrieves the introspection schema for a GraphQL API.

fn get_resolver(
&self,
input: GetResolverRequest
) -> RusotoFuture<GetResolverResponse, GetResolverError>

Retrieves a Resolver object.

fn get_schema_creation_status(
&self,
input: GetSchemaCreationStatusRequest
) -> RusotoFuture<GetSchemaCreationStatusResponse, GetSchemaCreationStatusError>

Retrieves the current status of a schema creation operation.

fn get_type(
&self,
input: GetTypeRequest
) -> RusotoFuture<GetTypeResponse, GetTypeError>

Retrieves a Type object.

fn list_api_keys(
&self,
input: ListApiKeysRequest
) -> RusotoFuture<ListApiKeysResponse, ListApiKeysError>

Lists the API keys for a given API.

API keys are deleted automatically sometime after they expire. However, they may still be included in the response until they have actually been deleted. You can safely call DeleteApiKey to manually delete a key before it's automatically deleted.

fn list_data_sources(
&self,
input: ListDataSourcesRequest
) -> RusotoFuture<ListDataSourcesResponse, ListDataSourcesError>

Lists the data sources for a given API.

fn list_graphql_apis(
&self,
input: ListGraphqlApisRequest
) -> RusotoFuture<ListGraphqlApisResponse, ListGraphqlApisError>

Lists your GraphQL APIs.

fn list_resolvers(
&self,
input: ListResolversRequest
) -> RusotoFuture<ListResolversResponse, ListResolversError>

Lists the resolvers for a given API and type.

fn list_types(
&self,
input: ListTypesRequest
) -> RusotoFuture<ListTypesResponse, ListTypesError>

Lists the types for a given API.

fn start_schema_creation(
&self,
input: StartSchemaCreationRequest
) -> RusotoFuture<StartSchemaCreationResponse, StartSchemaCreationError>

Adds a new schema to your GraphQL API.

This operation is asynchronous. Use to determine when it has completed.

fn update_api_key(
&self,
input: UpdateApiKeyRequest
) -> RusotoFuture<UpdateApiKeyResponse, UpdateApiKeyError>

Updates an API key.

fn update_data_source(
&self,
input: UpdateDataSourceRequest
) -> RusotoFuture<UpdateDataSourceResponse, UpdateDataSourceError>

Updates a DataSource object.

fn update_graphql_api(
&self,
input: UpdateGraphqlApiRequest
) -> RusotoFuture<UpdateGraphqlApiResponse, UpdateGraphqlApiError>

Updates a GraphqlApi object.

fn update_resolver(
&self,
input: UpdateResolverRequest
) -> RusotoFuture<UpdateResolverResponse, UpdateResolverError>

Updates a Resolver object.

fn update_type(
&self,
input: UpdateTypeRequest
) -> RusotoFuture<UpdateTypeResponse, UpdateTypeError>

Updates a Type object.