Crate rusoto_appsync

AWS AppSync provides API actions for creating and interacting with data sources using GraphQL from your application.

If you're using the service, you're probably looking for AppSyncClient and AppSync.

Structs

AdditionalAuthenticationProvider
Describes an additional authentication provider.

ApiKey

Describes an API key.

Customers invoke AWS AppSync GraphQL API operations with API keys as an identity mechanism. There are two key versions:

da1: This version was introduced at launch in November 2017. These keys always expire after 7 days. Key expiration is managed by Amazon DynamoDB TTL. The keys ceased to be valid after February 21, 2018 and should not be used after that date.

ListApiKeys returns the expiration time in milliseconds.

CreateApiKey returns the expiration time in milliseconds.

UpdateApiKey is not available for this key version.

DeleteApiKey deletes the item from the table.

Expiration is stored in Amazon DynamoDB as milliseconds. This results in a bug where keys are not automatically deleted because DynamoDB expects the TTL to be stored in seconds. As a one-time action, we will delete these keys from the table after February 21, 2018.

da2: This version was introduced in February 2018 when AppSync added support to extend key expiration.

ListApiKeys returns the expiration time in seconds.

CreateApiKey returns the expiration time in seconds and accepts a user-provided expiration time in seconds.

UpdateApiKey returns the expiration time in seconds and accepts a user-provided expiration time in seconds. Key expiration can only be updated while the key has not expired.

DeleteApiKey deletes the item from the table.

Expiration is stored in Amazon DynamoDB as seconds.

AppSyncClient
A client for the AWSAppSync API.

AuthorizationConfig
The authorization config in case the HTTP endpoint requires authorization.

AwsIamConfig
The AWS IAM configuration.

CognitoUserPoolConfig
Describes an Amazon Cognito user pool configuration.

CreateApiKeyRequest

CreateApiKeyResponse

CreateDataSourceRequest

CreateDataSourceResponse

CreateFunctionRequest

CreateFunctionResponse

CreateGraphqlApiRequest

CreateGraphqlApiResponse

CreateResolverRequest

CreateResolverResponse

CreateTypeRequest

CreateTypeResponse

DataSource
Describes a data source.

DeleteApiKeyRequest

DeleteApiKeyResponse

DeleteDataSourceRequest

DeleteDataSourceResponse

DeleteFunctionRequest

DeleteFunctionResponse

DeleteGraphqlApiRequest

DeleteGraphqlApiResponse

DeleteResolverRequest

DeleteResolverResponse

DeleteTypeRequest

DeleteTypeResponse

DynamodbDataSourceConfig
Describes an Amazon DynamoDB data source configuration.

ElasticsearchDataSourceConfig
Describes an Elasticsearch data source configuration.

FunctionConfiguration
A function is a reusable entity. Multiple functions can be used to compose the resolver logic.

GetDataSourceRequest

GetDataSourceResponse

GetFunctionRequest

GetFunctionResponse

GetGraphqlApiRequest

GetGraphqlApiResponse

GetIntrospectionSchemaRequest

GetIntrospectionSchemaResponse

GetResolverRequest

GetResolverResponse

GetSchemaCreationStatusRequest

GetSchemaCreationStatusResponse

GetTypeRequest

GetTypeResponse

GraphqlApi
Describes a GraphQL API.

HttpDataSourceConfig
Describes an HTTP data source configuration.

LambdaDataSourceConfig
Describes an AWS Lambda data source configuration.

ListApiKeysRequest

ListApiKeysResponse

ListDataSourcesRequest

ListDataSourcesResponse

ListFunctionsRequest

ListFunctionsResponse

ListGraphqlApisRequest

ListGraphqlApisResponse

ListResolversByFunctionRequest

ListResolversByFunctionResponse

ListResolversRequest

ListResolversResponse

ListTagsForResourceRequest

ListTagsForResourceResponse

ListTypesRequest

ListTypesResponse

LogConfig
The CloudWatch Logs configuration.

OpenIDConnectConfig
Describes an OpenID Connect configuration.

PipelineConfig
The pipeline configuration for a resolver of kind PIPELINE.

RdsHttpEndpointConfig
The Amazon RDS HTTP endpoint configuration.

RelationalDatabaseDataSourceConfig
Describes a relational database data source configuration.

Resolver
Describes a resolver.

StartSchemaCreationRequest

StartSchemaCreationResponse

TagResourceRequest

TagResourceResponse

Type
Describes a type.

UntagResourceRequest

UntagResourceResponse

UpdateApiKeyRequest

UpdateApiKeyResponse

UpdateDataSourceRequest

UpdateDataSourceResponse

UpdateFunctionRequest

UpdateFunctionResponse

UpdateGraphqlApiRequest

UpdateGraphqlApiResponse

UpdateResolverRequest

UpdateResolverResponse

UpdateTypeRequest

UpdateTypeResponse

UserPoolConfig
Describes an Amazon Cognito user pool configuration.

Enums

CreateApiKeyError
Errors returned by CreateApiKey

CreateDataSourceError
Errors returned by CreateDataSource

CreateFunctionError
Errors returned by CreateFunction

CreateGraphqlApiError
Errors returned by CreateGraphqlApi

CreateResolverError
Errors returned by CreateResolver

CreateTypeError
Errors returned by CreateType

DeleteApiKeyError
Errors returned by DeleteApiKey

DeleteDataSourceError
Errors returned by DeleteDataSource

DeleteFunctionError
Errors returned by DeleteFunction

DeleteGraphqlApiError
Errors returned by DeleteGraphqlApi

DeleteResolverError
Errors returned by DeleteResolver

DeleteTypeError
Errors returned by DeleteType

GetDataSourceError
Errors returned by GetDataSource

GetFunctionError
Errors returned by GetFunction

GetGraphqlApiError
Errors returned by GetGraphqlApi

GetIntrospectionSchemaError
Errors returned by GetIntrospectionSchema

GetResolverError
Errors returned by GetResolver

GetSchemaCreationStatusError
Errors returned by GetSchemaCreationStatus

GetTypeError
Errors returned by GetType

ListApiKeysError
Errors returned by ListApiKeys

ListDataSourcesError
Errors returned by ListDataSources

ListFunctionsError
Errors returned by ListFunctions

ListGraphqlApisError
Errors returned by ListGraphqlApis

ListResolversByFunctionError
Errors returned by ListResolversByFunction

ListResolversError
Errors returned by ListResolvers

ListTagsForResourceError
Errors returned by ListTagsForResource

ListTypesError
Errors returned by ListTypes

StartSchemaCreationError
Errors returned by StartSchemaCreation

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateApiKeyError
Errors returned by UpdateApiKey

UpdateDataSourceError
Errors returned by UpdateDataSource

UpdateFunctionError
Errors returned by UpdateFunction

UpdateGraphqlApiError
Errors returned by UpdateGraphqlApi

UpdateResolverError
Errors returned by UpdateResolver

UpdateTypeError
Errors returned by UpdateType

Traits

AppSync
Trait representing the capabilities of the AWSAppSync API. AWSAppSync clients implement this trait.