Struct rusoto_appsync::AppSyncClient

pub struct AppSyncClient { /* fields omitted */ }

A client for the AWSAppSync API.

Methods

impl AppSyncClient

pub fn new(region: Region) -> AppSyncClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> AppSyncClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,